Modularize `tables` module into one per database table
This patch doesn't change any functionality but just moves code
around.
